applespi: Preserve keyboard backlight level across reboots.
While desktop environments will also do this, this restores the
backlight level during early boot when, for example, the disk password
is being queried for.

Note that while Apple stores various settings in the EFI vars, they do
not store the keyboard backlight level there (and hence don't restore it
till the desktop comes up). I.e. there is no existing EFI var to use to
save this value, so this creates/uses a new one.